test(daemon): add extra tests for daemon

+++ test/cli/daemon.test.ts
@@ -8,7 +8,6 @@ import { path as ipfsExePathFunc } from "kubo";
 import dns from "node:dns";
 dns.setDefaultResultOrder("ipv4first"); // to be able to resolve localhost
 
-
 //@ts-ignore
 import chaiAsPromised from "chai-as-promised";
 chai.use(chaiAsPromised);
@@ -16,6 +15,16 @@ chai.use(chaiAsPromised);
 const { expect, assert } = chai;
 
 const rpcServerEndPoint = `ws://localhost:${defaults.PLEBBIT_RPC_API_PORT}`;
+
+const makeRequestToIpfsApi = async (apiPort: number) => {
+    return fetch(`http://localhost:${apiPort}/api/v0/bitswap/stat`, { method: "POST" });
+};
+
+const testConnectionToPlebbitRpc = async (rpcServerPort: number) => {
+    const rpcClient = new WebSocket(`ws://localhost:${rpcServerPort}`);
+    await new Promise((resolve) => setTimeout(resolve, 1000));
+    expect(rpcClient.readyState).to.equal(1); // 1 = connected
+};
 
 const startPlebbitDaemon = (args: string[]): Promise<ChildProcess> => {
     return new Promise(async (resolve, reject) => {
@@ -25,7 +34,7 @@ const startPlebbitDaemon = (args: string[]): Promise<ChildProcess> => {
             reject(`spawnAsync process '${daemonProcess.pid}' exited with code '${exitCode}' signal '${signal}'`);
         });
         daemonProcess.stdout.on("data", (data) => {
-            if (data.toString().match("Plebbit data path")) {
+            if (data.toString().match("Subplebbits in data path")) {
                 daemonProcess.removeAllListeners();
                 resolve(daemonProcess);
             }
@@ -55,7 +64,7 @@ const startIpfsDaemon = (args: string[]): Promise<ChildProcess> => {
     });
 };
 
-describe("plebbit daemon (ipfs daemon is started by plebbit)", async () => {
+describe("plebbit daemon (ipfs daemon is started by plebbit-cli)", async () => {
     let daemonProcess: ChildProcess;
 
     before(async () => {
@@ -124,10 +133,12 @@ describe(`plebbit daemon (ipfs daemon is started by another process on the same 
     before(async () => {
         await new Promise((resolve) => setTimeout(resolve, 5000)); // wait until the previous daemon is killed
         ipfsDaemonProcess = await startIpfsDaemon(["--init"]); // will start a daemon at 5001
+        const res = await makeRequestToIpfsApi(defaults.IPFS_API_PORT);
+        expect(res.status).to.equal(200);
     });
 
     after(async () => {
-        ipfsDaemonProcess.kill();
+        if (!ipfsDaemonProcess.killed) ipfsDaemonProcess.kill();
     });
 
     it(`plebbit daemon can use an ipfs node started by another program`, async () => {
@@ -137,5 +148,46 @@ describe(`plebbit daemon (ipfs daemon is started by another process on the same 
         expect(rpcClient.readyState).to.equal(1); // 1 = connected
         rpcClient.close();
         plebbitDaemonProcess.kill();
+    });
+
+    it(`plebbit daemon monitors ipfs API started by another process, and start a new IPFS process if needed`, async () => {
+        const plebbitDaemonProcess = await startPlebbitDaemon(["--plebbitDataPath", randomDirectory()]); // should use ipfsDaemonProcess
+        const rpcClient = new WebSocket(rpcServerEndPoint);
+        await new Promise((resolve) => setTimeout(resolve, 1000));
+        expect(rpcClient.readyState).to.equal(1); // 1 = connected
+
+        ipfsDaemonProcess.kill();
+        await new Promise((resolve) => setTimeout(resolve, 6000)); // plebbit daemon should start a new ipfs within 5 seconds ish
+        const resAfterRestart = await makeRequestToIpfsApi(defaults.IPFS_API_PORT);
+        expect(resAfterRestart.status).to.equal(200);
+
+        await plebbitDaemonProcess.kill();
+    });
+});
+
+describe(`plebbit daemon (relying on plebbit RPC started by another process)`, async () => {
+    let rpcProcess: ChildProcess;
+    before(async () => {
+        await new Promise((resolve) => setTimeout(resolve, 5000)); // wait until the previous daemon is killed
+        rpcProcess = await startPlebbitDaemon(["--plebbitDataPath", randomDirectory()]); // will start a daemon at 5001
+        await testConnectionToPlebbitRpc(defaults.PLEBBIT_RPC_API_PORT);
+    });
+
+    after(async () => {
+        if (!rpcProcess.killed) await rpcProcess.kill();
+    });
+
+    it(`plebbit daemon detects and uses another process' plebbit RPC`, async () => {
+        const anotherRpcProcess = await startPlebbitDaemon([]); // should start with no problem and use rpcProcess
+        await testConnectionToPlebbitRpc(defaults.PLEBBIT_RPC_API_PORT);
+        await anotherRpcProcess.kill(); // should not affect rpcProcess
+        await testConnectionToPlebbitRpc(defaults.PLEBBIT_RPC_API_PORT);
+    });
+    it(`plebbit daemon is monitoring another process' plebbit RPC and make sure it's always up`, async () => {
+        const anotherRpcProcess = await startPlebbitDaemon(["--plebbitDataPath", randomDirectory()]); // should monitor rpcProcess
+        rpcProcess.kill();
+        await new Promise((resolve) => setTimeout(resolve, 6000)); // wait until anotherRpcProcess restart the rpc
+        await testConnectionToPlebbitRpc(defaults.PLEBBIT_RPC_API_PORT);
+        await anotherRpcProcess.kill();
     });
 });
